feat: refactor stepper implementation and update documentation for clarity
